updated lesson 2  (#63)
* moving coding intro into coding-101

* js-core lesson 1 updated based on Node Girls workshop #6
The Node Girls workshop is a great immersive start to JS: I've amended some bits to use more accurate terminology, I also updated some of the code samples to be more relevant to our students. I changed the Array explanation to use Tim's explanation as it is easier to grasp. I've also moved the existing content of Lesson 1 to Lesson 2 as it will serve as a good way to reiterate the more complex concepts like arrays, objects etc...

* moving arrays and object from lesson 1 to lesson 2

* update JS-Core lesson 2